dynamic level instantiation in factory and powerups moved to new package
