+++ src/main/java/powerups/FreezePowerup.java
@@ -1,18 +1,19 @@
-package logic;
+package powerups;
+
 
 /**
  * A powerup which will let all circles stand still.
  * @author Bart
  *
  */
-public class SlowPowerup extends SpeedPowerup {
+public class FreezePowerup extends SpeedPowerup {
 	
-	private static final float MULTIPLIER = 0.5f;
+	private static final int MULTIPLIER = 0;
 	
 	/**
 	 * constructor.
 	 */
-	public SlowPowerup() {
+	public FreezePowerup() {
 		super(MULTIPLIER);
 	}
